+++ sdk/trace/tracer.go
@@ -7,7 +7,14 @@ import (
 	"context"
 	"time"
 
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/sdk"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
+	"go.opentelemetry.io/otel/sdk/trace/internal/x"
+	semconv "go.opentelemetry.io/otel/semconv/v1.36.0"
+	"go.opentelemetry.io/otel/semconv/v1.36.0/otelconv"
 	"go.opentelemetry.io/otel/trace"
 	"go.opentelemetry.io/otel/trace/embedded"
 )
@@ -17,9 +24,33 @@ type tracer struct {
 
 	provider             *TracerProvider
 	instrumentationScope instrumentation.Scope
+
+	selfObservabilityEnabled bool
+	spanLiveMetric           otelconv.SDKSpanLive
+	spanStartedMetric        otelconv.SDKSpanStarted
 }
 
 var _ trace.Tracer = &tracer{}
+
+func (tr *tracer) initSelfObservability() {
+	if !x.SelfObservability.Enabled() {
+		return
+	}
+
+	tr.selfObservabilityEnabled = true
+	mp := otel.GetMeterProvider()
+	m := mp.Meter("go.opentelemetry.io/otel/sdk/trace",
+		metric.WithInstrumentationVersion(sdk.Version()),
+		metric.WithSchemaURL(semconv.SchemaURL))
+
+	var err error
+	if tr.spanLiveMetric, err = otelconv.NewSDKSpanLive(m); err != nil {
+		otel.Handle(err)
+	}
+	if tr.spanStartedMetric, err = otelconv.NewSDKSpanStarted(m); err != nil {
+		otel.Handle(err)
+	}
+}
 
 // Start starts a Span and returns it along with a context containing it.
 //
@@ -46,6 +77,34 @@ func (tr *tracer) Start(
 	}
 
 	s := tr.newSpan(ctx, name, &config)
+	if tr.selfObservabilityEnabled {
+		// Check if the span has a parent span and set the origin attribute accordingly.
+		var attrParentOrigin attribute.KeyValue
+		if psc := trace.SpanContextFromContext(ctx); psc.IsValid() {
+			if psc.IsRemote() {
+				attrParentOrigin = tr.spanStartedMetric.AttrSpanParentOrigin(otelconv.SpanParentOriginRemote)
+			} else {
+				attrParentOrigin = tr.spanStartedMetric.AttrSpanParentOrigin(otelconv.SpanParentOriginLocal)
+			}
+		} else {
+			attrParentOrigin = tr.spanStartedMetric.AttrSpanParentOrigin(otelconv.SpanParentOriginNone)
+		}
+
+		// Determine the sampling result and create the corresponding attribute.
+		var attrSamplingResult attribute.KeyValue
+		if s.SpanContext().IsSampled() && s.IsRecording() {
+			attrSamplingResult = tr.spanStartedMetric.AttrSpanSamplingResult(
+				otelconv.SpanSamplingResultRecordAndSample,
+			)
+		} else if s.IsRecording() {
+			attrSamplingResult = tr.spanStartedMetric.AttrSpanSamplingResult(otelconv.SpanSamplingResultRecordOnly)
+		} else {
+			attrSamplingResult = tr.spanStartedMetric.AttrSpanSamplingResult(otelconv.SpanSamplingResultDrop)
+		}
+
+		tr.spanStartedMetric.Add(context.Background(), 1, attrParentOrigin, attrSamplingResult)
+	}
+
 	if rw, ok := s.(ReadWriteSpan); ok && s.IsRecording() {
 		sps := tr.provider.getSpanProcessors()
 		for _, sp := range sps {
@@ -152,6 +211,20 @@ func (tr *tracer) newRecordingSpan(
 
 	s.SetAttributes(sr.Attributes...)
 	s.SetAttributes(config.Attributes()...)
+
+	if tr.selfObservabilityEnabled {
+		// Determine the sampling result and create the corresponding attribute.
+		var attrSamplingResult attribute.KeyValue
+		if s.spanContext.IsSampled() {
+			attrSamplingResult = tr.spanLiveMetric.AttrSpanSamplingResult(
+				otelconv.SpanSamplingResultRecordAndSample,
+			)
+		} else {
+			attrSamplingResult = tr.spanLiveMetric.AttrSpanSamplingResult(otelconv.SpanSamplingResultRecordOnly)
+		}
+
+		tr.spanLiveMetric.Add(context.Background(), 1, attrSamplingResult)
+	}
 
 	return s
 }